+++ src/sc_client/client/_response_processor.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from sc_client import session
-from sc_client.constants import common
+from sc_client.constants import common as c
 from sc_client.constants.sc_types import ScType
 from sc_client.models import Response, ScAddr, ScEvent, ScLinkContent, ScLinkContentType, ScTemplateResult
 
@@ -16,42 +16,42 @@ class BaseResponseProcessor:
 
 class CreateElementsResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *_) -> list[ScAddr]:
-        return [ScAddr(addr_value) for addr_value in response.get(common.PAYLOAD)]
+        return [ScAddr(addr_value) for addr_value in response.get(c.PAYLOAD)]
 
 
 class CreateElementsBySCsResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *_) -> list[bool]:
-        return [bool(result) for result in response.get(common.PAYLOAD)]
+        return [bool(result) for result in response.get(c.PAYLOAD)]
 
 
 class CheckElementsResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *_) -> list[ScType]:
-        return [ScType(type_value) for type_value in response.get(common.PAYLOAD)]
+        return [ScType(type_value) for type_value in response.get(c.PAYLOAD)]
 
 
 class DeleteElementsResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *_) -> bool:
-        return response.get(common.STATUS)
+        return response.get(c.STATUS)
 
 
 class SetLinkContentResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *_) -> bool:
-        return response.get(common.STATUS)
+        return response.get(c.STATUS)
 
 
 class GetLinkContentResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *_) -> list[ScLinkContent]:
-        response_payload = response.get(common.PAYLOAD)
+        response_payload = response.get(c.PAYLOAD)
         result = []
         for link in response_payload:
-            str_type: str = link.get(common.TYPE)
-            result.append(ScLinkContent(link.get(common.VALUE), ScLinkContentType[str_type.upper()]))
+            str_type: str = link.get(c.TYPE)
+            result.append(ScLinkContent(link.get(c.VALUE), ScLinkContentType[str_type.upper()]))
         return result
 
 
 class GetLinksByContentResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *_) -> list[list[ScAddr]]:
-        response_payload = response.get(common.PAYLOAD)
+        response_payload = response.get(c.PAYLOAD)
         if response_payload:
             return [[ScAddr(addr_value) for addr_value in addr_list] for addr_list in response_payload]
         return response_payload
@@ -61,9 +61,15 @@ class GetLinksByContentSubstringResponseProcessor(GetLinksByContentResponseProce
     pass
 
 
+class GetLinksContentsByContentSubstringResponseProcessor(BaseResponseProcessor):
+    def __call__(self, response: Response, *_) -> list[list[ScAddr]]:
+        response_payload = response.get(c.PAYLOAD)
+        return response_payload
+
+
 class ResolveKeynodesResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *_) -> list[ScAddr]:
-        response_payload = response.get(common.PAYLOAD)
+        response_payload = response.get(c.PAYLOAD)
         if response_payload:
             return [ScAddr(addr_value) for addr_value in response_payload]
         return response
@@ -72,10 +78,10 @@ class ResolveKeynodesResponseProcessor(BaseResponseProcessor):
 class TemplateSearchResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *_) -> list[ScTemplateResult]:
         result = []
-        if response.get(common.STATUS):
-            response_payload = response.get(common.PAYLOAD)
-            aliases = response_payload.get(common.ALIASES)
-            all_addrs = response_payload.get(common.ADDRS)
+        if response.get(c.STATUS):
+            response_payload = response.get(c.PAYLOAD)
+            aliases = response_payload.get(c.ALIASES)
+            all_addrs = response_payload.get(c.ADDRS)
             for addrs_list in all_addrs:
                 addrs = [ScAddr(addr) for addr in addrs_list]
                 result.append(ScTemplateResult(addrs, aliases))
@@ -85,10 +91,10 @@ class TemplateSearchResponseProcessor(BaseResponseProcessor):
 class TemplateGenerateResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *_) -> ScTemplateResult:
         result = None
-        if response.get(common.STATUS):
-            response_payload = response.get(common.PAYLOAD)
-            aliases = response_payload.get(common.ALIASES)
-            addrs_list = response_payload.get(common.ADDRS)
+        if response.get(c.STATUS):
+            response_payload = response.get(c.PAYLOAD)
+            aliases = response_payload.get(c.ALIASES)
+            addrs_list = response_payload.get(c.ADDRS)
             addrs = [ScAddr(addr) for addr in addrs_list]
             result = ScTemplateResult(addrs, aliases)
         return result
@@ -98,7 +104,7 @@ class EventsCreateResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *events: ScEvent) -> list[ScEvent]:
         result = []
         for count, event in enumerate(events):
-            command_id = response.get(common.PAYLOAD)[count]
+            command_id = response.get(c.PAYLOAD)[count]
             sc_event = ScEvent(command_id, event.event_type, event.callback)
             session.set_event(sc_event)
             result.append(sc_event)
@@ -109,26 +115,27 @@ class EventsDestroyResponseProcessor(BaseResponseProcessor):
     def __call__(self, response: Response, *events: ScEvent) -> bool:
         for event in events:
             session.drop_event(event.id)
-        return response.get(common.STATUS)
+        return response.get(c.STATUS)
 
 
 class ResponseProcessor:
     _response_request_mapper = {
-        common.ClientCommand.CREATE_ELEMENTS: CreateElementsResponseProcessor(),
-        common.ClientCommand.CREATE_ELEMENTS_BY_SCS: CreateElementsBySCsResponseProcessor(),
-        common.ClientCommand.CHECK_ELEMENTS: CheckElementsResponseProcessor(),
-        common.ClientCommand.DELETE_ELEMENTS: DeleteElementsResponseProcessor(),
-        common.ClientCommand.KEYNODES: ResolveKeynodesResponseProcessor(),
-        common.ClientCommand.GET_LINK_CONTENT: GetLinkContentResponseProcessor(),
-        common.ClientCommand.GET_LINKS_BY_CONTENT: GetLinksByContentResponseProcessor(),
-        common.ClientCommand.GET_LINKS_BY_CONTENT_SUBSTRING: GetLinksByContentSubstringResponseProcessor(),
-        common.ClientCommand.SET_LINK_CONTENTS: SetLinkContentResponseProcessor(),
-        common.ClientCommand.EVENTS_CREATE: EventsCreateResponseProcessor(),
-        common.ClientCommand.EVENTS_DESTROY: EventsDestroyResponseProcessor(),
-        common.ClientCommand.GENERATE_TEMPLATE: TemplateGenerateResponseProcessor(),
-        common.ClientCommand.SEARCH_TEMPLATE: TemplateSearchResponseProcessor(),
+        c.ClientCommand.CREATE_ELEMENTS: CreateElementsResponseProcessor(),
+        c.ClientCommand.CREATE_ELEMENTS_BY_SCS: CreateElementsBySCsResponseProcessor(),
+        c.ClientCommand.CHECK_ELEMENTS: CheckElementsResponseProcessor(),
+        c.ClientCommand.DELETE_ELEMENTS: DeleteElementsResponseProcessor(),
+        c.ClientCommand.KEYNODES: ResolveKeynodesResponseProcessor(),
+        c.ClientCommand.GET_LINK_CONTENT: GetLinkContentResponseProcessor(),
+        c.ClientCommand.GET_LINKS_BY_CONTENT: GetLinksByContentResponseProcessor(),
+        c.ClientCommand.GET_LINKS_BY_CONTENT_SUBSTRING: GetLinksByContentSubstringResponseProcessor(),
+        c.ClientCommand.GET_LINKS_CONTENTS_BY_CONTENT_SUBSTRING: GetLinksContentsByContentSubstringResponseProcessor(),
+        c.ClientCommand.SET_LINK_CONTENTS: SetLinkContentResponseProcessor(),
+        c.ClientCommand.EVENTS_CREATE: EventsCreateResponseProcessor(),
+        c.ClientCommand.EVENTS_DESTROY: EventsDestroyResponseProcessor(),
+        c.ClientCommand.GENERATE_TEMPLATE: TemplateGenerateResponseProcessor(),
+        c.ClientCommand.SEARCH_TEMPLATE: TemplateSearchResponseProcessor(),
     }
 
-    def run(self, request_type: common.ClientCommand, *args, **kwargs):
+    def run(self, request_type: c.ClientCommand, *args, **kwargs):
         response_processor = self._response_request_mapper.get(request_type)
         return response_processor(*args, **kwargs)